This is the first version of source code

+++ src/main/java/com/zhineng/test/dao/OrganizationDao.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrganizationDao extends JpaRepository<Organization, Long> {}
+public interface OrganizationDao extends JpaRepository<Organization, Integer> {}
